requirements.txt
Django==4.2.2
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pillow==11.3.0
black==25.1.0
isort==6.0.1
django-filter==25.1
djangorestframework_simplejwt==5.5.1
ipython==7.34.0
drf-yasg==1.21.10
stripe==12.5.0
forex-python==1.9.2
celery==5.5.3
redis==6.2.0
django-celery-beat==2.8.1
django-redis
gunicorn
